Added TextBox, rework Token system
Added IToken interface to have multiple types of token as opposed to tokens with different kinds of tags.
Created TextBox class, which is a box that is now capable of drawing tokens within its limits and will be handling the links and images in the future.
The DrawablePage now uses a TextBox instead of drawing the text by itself.
Known problems:
- Page changes not working as intended, page change system will need rework.